pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Sokolov Fedor <sokol_f4@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.11"
alembic = "1.13.0"
fastapi = "0.75.1"
fastapi-users = "12.1.2"
loguru = "0.7.2"
asyncpg = "0.27.0"
python-dotenv = "1.0.0"
sqlalchemy = "2.0.23"
uvicorn = "0.24.0.post1"
fastapi-users-db-sqlalchemy = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "22.10.0"
isort = "5.12.0"
pylint = "3.0.2"
mypy = "1.7.1"
pre-commit = "^3.5.0"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
